+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## LibFrameFade
+### v3
+- Fixed an issue where ElvUI's chat frame tabs would incorrectly fade out when not actively hovered over.
+
 ### v2
 - Fixed an issue where animations could have incorrect durations in edge cases where the frames had already been partially animated by the default UIFrameFade manager.
 - Fixed an issue where user code that modifies the on-finish callback or its arguments in the fadeInfo table directly would not result in the correct function or arguments being used when the animation finishes.